+++ CHANGELOG.md
@@ -12,6 +12,19 @@ This is the changelog for Cldr_Dates_Times v2.9.0 released on August 14th, 2021.
 
 * Allow formatting of intervals where one side is `nil`. This will produce an open-ended interval. Only one side of the interval can be `nil`. Thanks to @woylie for the request.  Closes #23.
 
+#### Examples
+
+      iex> Cldr.Date.Interval.to_string ~D[2020-01-01], nil, MyApp.Cldr,
+      ...> format: :short
+      {:ok, "1/1/20 –"}
+
+      iex> Cldr.Time.Interval.to_string ~U[2020-01-01 00:00:00.0Z], nil, MyApp.Cldr,
+      ...> format: :long, style: :flex
+      {:ok, "12:00:00 AM UTC –"}
+
+      iex> Cldr.DateTime.Interval.to_string ~U[2020-01-01 00:00:00.0Z], nil, MyApp.Cldr
+      {:ok, "Jan 1, 2020, 12:00:00 AM –"}
+
 ## Cldr_Dates_Times v2.8.0
 
 This is the changelog for Cldr_Dates_Times v2.8.0 released on July 1st, 2021.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_cldr_dates_times/tags)